pyproject.toml
[project]
name = "autodraft-ingest"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.2.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]
